+++ app/src/main/jni/breathing/breathing.c
@@ -34,7 +34,7 @@ void initialise_breathing_buffer(BreathingBuffer *breathing_buffer) {
     initialise_activity_level_buffer(&activity_level_buffer);
 }
 
-void update_breathing(double *new_accel_data_original, BreathingBuffer *breathing_buffer) {
+void update_breathing(float *new_accel_data_original, BreathingBuffer *breathing_buffer) {
     // We assume apriori that the current breathing value is not valid.
     breathing_buffer->is_valid = false;
 
@@ -43,12 +43,13 @@ void update_breathing(double *new_accel_data_original, BreathingBuffer *breathin
     }
 
     // Return if any of the acceleration values are NAN. This shouldn't happen.
-    if (isnan(new_accel_data_original[0]) || isnan(new_accel_data_original[1]) || isnan(new_accel_data_original[2])) {
+    if (isnan((float) new_accel_data_original[0]) || isnan((float) new_accel_data_original[1]) ||
+        isnan((float) new_accel_data_original[2])) {
         return;
     }
 
     // First, make a copy of input accel vector and use that in the following
-    double new_accel_data[3];
+    float new_accel_data[3];
     copy_accel_vector(new_accel_data, new_accel_data_original);
 
     // Save the position where the next activity level will be saved. This is the current position in the buffer.
@@ -105,15 +106,15 @@ void update_breathing(double *new_accel_data_original, BreathingBuffer *breathin
     }
 
     // Breathing signal calculation
-    double final_bs = dot_product(rotation_axis_buffer.current_axis, mean_axis_buffer.mean_axis);
+    float final_bs = dot_product(rotation_axis_buffer.current_axis, mean_axis_buffer.mean_axis);
 //    __android_log_print(ANDROID_LOG_INFO, "BS", "bs: %lf", final_bs);
     // TODO: this should be completely unnecessary, as the amplitude of the breathing signal ins meaningless anyway
     final_bs = final_bs * SAMPLE_RATE * 10.0f;
 
     // Breathing angle calculation
-    double mean_accel_cross_mean_axis[3];
+    float mean_accel_cross_mean_axis[3];
     cross_product(mean_accel_cross_mean_axis, mean_unit_accel_buffer.mean_unit_vector, mean_axis_buffer.mean_axis);
-    double final_ba;
+    float final_ba;
     final_ba = dot_product(mean_accel_cross_mean_axis, new_accel_data);
 
     // Smooth the breathing signal and angles for the last time
