+++ app/src/main/jni/breathing/breathing_rate_stats.h
@@ -15,28 +15,28 @@
 
 typedef struct {
     int fill;
-    double breathing_rates[BREATHING_RATES_BUFFER_SIZE];
+    float breathing_rates[BREATHING_RATES_BUFFER_SIZE];
     bool is_valid;
-    double previous_mean;
-    double current_mean;
-    double previous_variance;
-    double current_variance;
-    double max;
-    double min;
+    float previous_mean;
+    float current_mean;
+    float previous_variance;
+    float current_variance;
+    float max;
+    float min;
 } BreathingRateBuffer;
 
 void initialise_breathing_rate_buffer(BreathingRateBuffer *breathing_rate_buffer);
 
-void update_breathing_rate_buffer(double breathing_rate, BreathingRateBuffer *breathing_rate_buffer);
+void update_breathing_rate_buffer(float breathing_rate, BreathingRateBuffer *breathing_rate_buffer);
 
 void calculate_breathing_rate_stats(BreathingRateBuffer *breathing_rate_buffer);
 
 int breathing_rate_number_of_breaths(BreathingRateBuffer *breathing_rate_buffer);
 
-double breathing_rate_mean(BreathingRateBuffer *breathing_rate_buffer);
+float breathing_rate_mean(BreathingRateBuffer *breathing_rate_buffer);
 
-double breathing_rate_variance(BreathingRateBuffer *breathing_rate_buffer);
+float breathing_rate_variance(BreathingRateBuffer *breathing_rate_buffer);
 
-double breathing_rate_standard_deviation(BreathingRateBuffer *breathing_rate_buffer);
+float breathing_rate_standard_deviation(BreathingRateBuffer *breathing_rate_buffer);
 
 #endif
