+++ app/src/main/jni/breathing/mean_axis_buffer.h
@@ -9,17 +9,17 @@
 typedef struct
 {
 
-	double sum[3];
+	float sum[3];
 	int current_position;
 	int fill;
 
-	double accel_buffer[MEAN_AXIS_SIZE][3];
-	double mean_axis[3];
+	float accel_buffer[MEAN_AXIS_SIZE][3];
+	float mean_axis[3];
 	bool is_valid;
 
 } MeanAxisBuffer;
 
 void initialise_mean_axis_buffer(MeanAxisBuffer *mean_axis_buffer);
-void update_mean_axis_buffer(double *new_accel_data, MeanAxisBuffer *mean_axis_buffer);
+void update_mean_axis_buffer(float *new_accel_data, MeanAxisBuffer *mean_axis_buffer);
 
 #endif
