+++ app/src/main/jni/breathing/mean_unit_accel_buffer.h
@@ -8,18 +8,18 @@
 
 typedef struct {
 
-    double sum[3];
+    float sum[3];
     int current_position;
     int fill;
 
-    double values[MEAN_ACCEL_BUFFER_SIZE][3];
-    double mean_unit_vector[3];
+    float values[MEAN_ACCEL_BUFFER_SIZE][3];
+    float mean_unit_vector[3];
     bool is_valid;
 
 } MeanUnitAccelBuffer;
 
 void initialise_mean_unit_accel_buffer(MeanUnitAccelBuffer *mean_accel_buffer);
 
-void update_mean_unit_accel_buffer(double *new_accel_data, MeanUnitAccelBuffer *mean_accel_buffer);
+void update_mean_unit_accel_buffer(float *new_accel_data, MeanUnitAccelBuffer *mean_accel_buffer);
 
 #endif
