+++ app/src/main/jni/breathing/rotation_axis.h
@@ -6,16 +6,16 @@
 
 typedef struct {
 
-    double previous_accel_data[3];
+    float previous_accel_data[3];
     bool is_previous_accel_data_valid;
 
-    double current_axis[3];
+    float current_axis[3];
     bool is_current_axis_valid;
 
 } RotationAxisBuffer;
 
 void initialise_rotation_axis_buffer(RotationAxisBuffer *axis_and_angle_buffer);
 
-void update_rotation_axis_buffer(double *new_accel_data, RotationAxisBuffer *rotation_axis_buffer);
+void update_rotation_axis_buffer(float *new_accel_data, RotationAxisBuffer *rotation_axis_buffer);
 
 #endif
